+++ source/include/sys/task/lock.h
@@ -35,7 +35,7 @@
  * @param flags flags (LOCK_*)
  * @return 0 on success
  */
-int lock_aquire(tPid pid,uint ident,ushort flags);
+int lock_aquire(tPid pid,ulong ident,ushort flags);
 
 /**
  * Releases the lock with given ident and pid
@@ -44,7 +44,7 @@ int lock_aquire(tPid pid,uint ident,ushort flags);
  * @param ident to identify the lock
  * @return 0 on success
  */
-int lock_release(tPid pid,uint ident);
+int lock_release(tPid pid,ulong ident);
 
 /**
  * Releases all locks of the given process
