+++ source/kernel/src/syscalls/other.c
@@ -67,7 +67,7 @@ void sysc_debug(sIntrptStackFrame *stack) {
 
 void sysc_getConf(sIntrptStackFrame *stack) {
 	uint id = SYSC_ARG1(stack);
-	int res = conf_get(id);
+	long res = conf_get(id);
 	if(res < 0)
 		SYSC_ERROR(stack,res);
 	SYSC_RET1(stack,res);
